Set libarchive to build position-independent
When building on x86_64, it would fail while attempting to link to a
shared object of the libarchive library. It would only find a static
library available. It doesn't appear that the position-independent flag
(-fPIC) was getting applied to the libarchive recipe. Add the
position-independent flag to the extra cmake arguments to fix the issue.